Add image index property to use for index variable
Once the file is saved, if the user undoes pasting an image, that image
file will be stored with an index calculating from the length of the
`pasted_images` dictionary, while the dictionary doesn't keep that image
(it only keeps if an image is in the `.blend` file). Later, if the user
pastes another image, its index will be the same as the previous image
(probably not the undone image) and it will be overwritten. This change
will prevent that by incrementing the image index independently of the
`pasted_images` dictionary.