Remove `annotations` statement from `__future__`
There is a bug that make class's annotation properties not working when
using `from __future__ import annotations` (in this module it is the
`timeout` property from the check update operator). Although the issue
doesn't exist in 'Blender' 2.93, we have to remove the `annotations`
statement to make it compilable with 'Blender' 2.83.

To read more about the bug, please refer to:
https://developer.blender.org/T63713.

The `annotations` statement is needed to support forward references. In
this module, it isn't necessary but it helps avoid the quoting of the
`list[bpy.types.Image]`. Removing it, we have to reverse them.

To read more about the statement, please refer to:
https://docs.python.org/3/whatsnew/3.7.html#pep-563-postponed-evaluation-of-annotations.