Copy a render result will save it to the `tempdir`
The copy operator in the 'Image Editor' will save the render result
images to the temporary 'Blender' directory instead of saving them to
the target save directory because the user doesn't intend to do so.

To save it to the save directory, they have to do manually or have to
copy then paste to mark it as a pasted image.